Refactor requests into specific package and rename transformer appropriately
