+++ src/main/java/uk/gov/companieshouse/web/emergencyauthcodeweb/model/emergencyauthcode/officer/EACOfficerList.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.web.emergencyauthcodeweb.model.emergencyauthcode.officer;
+
+public class EACOfficerList {
+}
