+++ src/main/java/uk/gov/companieshouse/web/emergencyauthcodeweb/transformer/emergencyauthcode/officer/EACOfficerListTransformer.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.web.emergencyauthcodeweb.transformer.emergencyauthcode.officer;
+
+public interface EACOfficerListTransformer {
+}
