+++ src/test/java/uk/gov/companieshouse/web/emergencyauthcodeweb/transformer/emergencyauthcode/officer/EACOfficerTransformerTest.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.web.emergencyauthcodeweb.transformer.emergencyauthcode.officer;
+
+public class EACOfficerTransformerTest {
+}
